+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added `DriverManager::get_drivers_for_filename` for the ability to auto detect compatible `Driver`s for writing data.
+
 - Added `Feature::unset_field`
   - <https://github.com/georust/gdal/pull/503>